Move API methods to Api class
https://bugzilla.redhat.com/show_bug.cgi?id=1369842
